Ai Projects/autonomous-business-intelligence/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.5
langchain==0.0.340
langchain-openai==0.0.2
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
prophet==1.1.5
statsmodels==0.14.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
pymongo==4.6.0
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
pytest==7.4.3
pytest-asyncio==0.21.1
schedule==1.2.0

